use base class method for Regression.conf_int, move glmnet_r_results to results, implement tsa.unit_root
