+++ sm2/tsa/vector_ar/tests/JMulTi_results/parse_jmulti_vecm_output.py
@@ -5,38 +5,7 @@ from io import open
 import itertools
 import numpy as np
 
-debug_mode = False
-
 cur_dir = os.path.dirname(os.path.realpath(__file__))
-
-
-def print_debug_output(results, dt):
-    print("\n\n\nDETERMINISTIC TERMS: " + dt)
-    alpha = results["est"]["alpha"]
-    print("alpha:")
-    print(str(type(alpha)) + str(alpha.shape))
-    print(alpha)
-    print("se: ")
-    print(results["se"]["alpha"])
-    print("t: ")
-    print(results["t"]["alpha"])
-    print("p: ")
-    print(results["p"]["alpha"])
-    beta = results["est"]["beta"]
-    print("beta:")
-    print(str(type(beta)) + str(beta.shape))
-    print(beta)
-    gamma = results["est"]["Gamma"]
-    print("Gamma:")
-    print(str(type(gamma)) + str(gamma.shape))
-    print(gamma)
-    if "co" in dt or "s" in dt or "lo" in dt:
-        c = results["est"]["C"]
-        print("C:")
-        print(str(type(c)) + str(c.shape))
-        print(c)
-        print("se: ")
-        print(results["se"]["C"])
 
 
 def dt_s_tup_to_string(dt_s_tup):
@@ -377,8 +346,6 @@ def load_results_jmulti(dataset):
         results["whiteness"] = wresults
 
         # ---------------------------------------------------------------------
-        if debug_mode:
-            print_debug_output(results, dt_string)
 
         results_dict_per_det_terms[dt_s] = results
